Added support for creating and listing models from cli
